Sponsor 컴포넌트 내 constants 파일 이동 및 PDF 다운로드 경로 로컬에서 S3로 변경
