--- components/sponsor/information/constants/images.ts
@@ -1,32 +1,0 @@
-import { IMAGE_BASE_URL } from '@/constants';
-
-export const SPONSOR_MEANING_1920 = `${IMAGE_BASE_URL}/sponsor_meaning_1920.jpg`;
-export const SPONSOR_MEANING_1280 = `${IMAGE_BASE_URL}/sponsor_meaning_1280.jpg`;
-export const SPONSOR_MEANING_360 = `${IMAGE_BASE_URL}/sponsor_meaning_360.jpg`;
-
-export const SPONSOR_SELECTION_PROCESS_1920 = `${IMAGE_BASE_URL}/sponsor_selection_process_1920.jpg`;
-export const SPONSOR_SELECTION_PROCESS_1280 = `${IMAGE_BASE_URL}/sponsor_selection_process_1280.jpg`;
-export const SPONSOR_SELECTION_PROCESS_360 = `${IMAGE_BASE_URL}/sponsor_selection_process_360.jpg`;
-
-export const SPONSOR_BOOTH_1920 = `${IMAGE_BASE_URL}/sponsor_booth_1920.jpg`;
-export const SPONSOR_BOOTH_1280 = `${IMAGE_BASE_URL}/sponsor_booth_1280.jpg`;
-export const SPONSOR_BOOTH_360 = `${IMAGE_BASE_URL}/sponsor_booth_360.jpg`;
-
-export const SPONSOR_GIFT_1920 = `${IMAGE_BASE_URL}/sponsor_gift_1920.jpg`;
-export const SPONSOR_GIFT_1280 = `${IMAGE_BASE_URL}/sponsor_gift_1280.jpg`;
-export const SPONSOR_GIFT_360 = `${IMAGE_BASE_URL}/sponsor_gift_360.jpg`;
-
-export const SPONSOR_LOGO_1920 = `${IMAGE_BASE_URL}/sponsor_logo_1920.jpg`;
-export const SPONSOR_LOGO_1280 = `${IMAGE_BASE_URL}/sponsor_logo_1280.jpg`;
-export const SPONSOR_LOGO_360 = `${IMAGE_BASE_URL}/sponsor_logo_360.jpg`;
-
-export const SPONSOR_SESSIONS_1920 = `${IMAGE_BASE_URL}/sponsor_sessions_1920.jpg`;
-export const SPONSOR_SESSIONS_1280 = `${IMAGE_BASE_URL}/sponsor_sessions_1280.jpg`;
-export const SPONSOR_SESSIONS_360 = `${IMAGE_BASE_URL}/sponsor_sessions_360.jpg`;
-
-export const SPONSORSHIP_CERTIFICATE_1920 = `${IMAGE_BASE_URL}/sponsorship_certificate_1920.jpg`;
-export const SPONSORSHIP_CERTIFICATE_1280 = `${IMAGE_BASE_URL}/sponsorship_certificate_1280.jpg`;
-export const SPONSORSHIP_CERTIFICATE_360 = `${IMAGE_BASE_URL}/sponsorship_certificate_360.jpg`;
-
-export const SPONSOR_VIDEO_1920 = `${IMAGE_BASE_URL}/sponsor_video_1920.jpg`;
-export const SPONSOR_TICKET_1920 = `${IMAGE_BASE_URL}/sponsor_ticket_1920.png`;
