+++ constants/index.ts
@@ -1,2 +1,5 @@
 export const IMAGE_BASE_URL =
   'https://pyconkr-web-2023-static.s3.ap-northeast-2.amazonaws.com';
+
+export const GUIDE_FILE_DOWNLOAD =
+  'https://pyconkr-web-2023-static.s3.ap-northeast-2.amazonaws.com/pyconkr_2023_sponsorship_guide.pdf';
